refactor: move Footer.astro component into a separate folder
