+++ src/components/footer/Footer.astro
@@ -1,6 +1,6 @@
 ---
-import type { RichTextContent } from "../lib/content/models/rich-text-block";
-import RichTextBlockContent from "./rich-text/RichTextBlock.astro";
+import type { RichTextContent } from "../../lib/content/models/rich-text-block";
+import RichTextBlockContent from "../rich-text/RichTextBlock.astro";
 
 interface Props {
 	content: RichTextContent;
