+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import Footer from "../components/Footer.astro";
+import Footer from "../components/footer/Footer.astro";
 import Header from "../components/header/Header.astro";
 import MetaData from "../components/meta-data/MetaData.astro";
 import { createContentProvider } from "../lib/content/provider";
